Add HAVE_SOCKLEN_T to config.h for FFmpeg
Needed to fix compilation after recent FFmpeg changes. It's now always
set to true without any tests. I don't expect this to cause problems
as common systems will have the type and the FFmpeg demuxers which
would use it are not compiled under MPlayer (compilation was broken
because the type was redefined in a header).


git-svn-id: svn://svn.mplayerhq.hu/mplayer/trunk@25592 b3059339-0415-0410-9bf9-f77b7e298cf2