Separate automatic dependency generation flags from general CFLAGS.
Before, there was an unfortunate interaction with 'make checkheaders':
Compiling a .h file would generate a .d dependency information file for
that .h file as a sideeffect of compilation. Unfortunately this would
clobber the .d files of the .c files with the same basename, resulting
in broken dependency information when running plain make.


git-svn-id: svn://svn.mplayerhq.hu/mplayer/trunk@31614 b3059339-0415-0410-9bf9-f77b7e298cf2