Prevent an out-of-bounds array access.
This out-of-bounds array access was of not much effect until r33190, because
the variables falsely changed (set to 1) were either almost always unused
(the last member of the mainItems array) or variable menuIsPresent which only
has effect if there is a menu present (in which case it is set anyway).

With r33190 the wItem (and thus the guiItems) structure size changed and now
variables were accessed whose changings have visible effect. After releasing
the moved main (or playbar) window, the IndexOfItems values got destroyed and
items disappeared.


git-svn-id: svn://svn.mplayerhq.hu/mplayer/trunk@33194 b3059339-0415-0410-9bf9-f77b7e298cf2