Rename sub window video window.
It was a bad idea to name the video window "sub window" at the time
the GUI was written. The term "sub window" does make sense from the
programmer's point of view, but it doesn't make any sense at all from
the user's point of view, because the sub window simply is the window
where the video will be displayed.

Moreover, since the term "sub" is generally short for "subtitles",
the renaming makes the code much easier to understand.


git-svn-id: svn://svn.mplayerhq.hu/mplayer/trunk@34785 b3059339-0415-0410-9bf9-f77b7e298cf2