Allow plain text playlists for the GUI.
This additional plain text playlist parsing will only happen if
the user selects from the "Playlists" filter of the file selector.

The "Playlists" filter is the GUI's equivalent to the -playlist option,
so there won't happen more bad things than with the option. (Actually,
it's safer than the option, because the "Playlists" filter only offers
a limited selection.)


git-svn-id: svn://svn.mplayerhq.hu/mplayer/trunk@34976 b3059339-0415-0410-9bf9-f77b7e298cf2