Internally map item 'potmeter' onto 'hpotmeter'.
Former version of the GUI treated a potmeter very similar to a hpotmeter
(the Win32 GUI still does so) and lots of skins are solely using
potmeters instead of hpotmeters, although this doesn't make sense at
all.

The current version of the GUI is treating a potmeter differently, but
in order to not break old skins, restore the old behaviour.

For the X11/GTK GUI, a potmeter is now simply a hpotmeter with
button=NULL and (button)width=(button)height=0. For the Win32 GUI
(where skins unfortunately are handled a bit differently and things
are more complicated) a potmeter is now a hpotmeter without button
but (button)width=(widget)width and (button)height=(widget)height.

Additionally, print a legacy information, because the item 'potmeter' is
obsolete now and oughtn't be used any longer.


git-svn-id: svn://svn.mplayerhq.hu/mplayer/trunk@36988 b3059339-0415-0410-9bf9-f77b7e298cf2