Pass visual information to OpenGL video output driver for X11.
The GUI and the OpenGL video output driver for X11 choose different
visuals which results in a BadMatch error from XChangeWindowAttributes()
in x11_common.c.

Since the GUI has to choose the visual first to create its main window,
it is necessary to tell it the drivers so that they can use this same
(hopefully suitable) visual.


git-svn-id: svn://svn.mplayerhq.hu/mplayer/trunk@38174 b3059339-0415-0410-9bf9-f77b7e298cf2