ad_hwmpa.c: avoid potential out of bounds.
This happens when the frame size is larger than the
size of the decoded audio - obviously this should not
happen in real audio files as it doesn't make sense
to select a compression mode that increases size.
Note that unless all other MPEG audio codecs are disabled
this codec will never be auto-selected.

Fixes trac issue #2357.

git-svn-id: svn://svn.mplayerhq.hu/mplayer/trunk@38224 b3059339-0415-0410-9bf9-f77b7e298cf2