Enable Native Language Support.
Add a configure option that allows specifying multiple languages for
messages (for the moment only for the GUI). By default, the option is
enabled for the GUI and disabled otherwise.

With option --disable-nls or when specifying only one language for
messages, nothing changes from the behavior so far, i.e. all texts
are compiled into the binary.

With option --enable-nls and specifying multiple languages, the English
texts are compiled into the binary and *.mo files are created for all
other languages.

For this, help_create_po.pl extracts the defined text constants from
the help/help_mp-*.h files and creates the necessary *.po intermediate
files.


git-svn-id: svn://svn.mplayerhq.hu/mplayer/trunk@38300 b3059339-0415-0410-9bf9-f77b7e298cf2