muxer_lavf: fix compilation with newer FFmpeg.
Note that this also moves the time base setting from codec to
stream, which makes it actually work again as the codec
time base is ignored by FFmpeg.
Leaves ve_lavc to fix for mencoder, which is a bit more complicated,
mostly because it's so many parameters that need to be replaced
with the matching dictionary setting, which is not easy to find.

git-svn-id: svn://svn.mplayerhq.hu/mplayer/trunk@38309 b3059339-0415-0410-9bf9-f77b7e298cf2