Realize canceling the file sector dialog through an accelerator group.
Let the accelerator group handle the escape key instead of handling
the key press event ourselves. Accelerator groups are already used
for this purpose in all other dialogs.

This way, the escape key is now also recognized everywhere in the file
selector dialog and not only in the file list.


git-svn-id: svn://svn.mplayerhq.hu/mplayer/trunk@38320 b3059339-0415-0410-9bf9-f77b7e298cf2