configure: Do not try to detect GnuTLS if OpenSSL was enabled
Enabling GnuTLS via auto detection could have us end up with both
libraries enabled. This won't build anymore. IIRC at some point it
did build, though it was not easy to figure out which implementation
would be used in the end.

This change is done in the light of having no auto detection for
OpenSSL and probably never will, because of the OpenSSL license.


git-svn-id: svn://svn.mplayerhq.hu/mplayer/trunk@38339 b3059339-0415-0410-9bf9-f77b7e298cf2