mplayer: uninit_player: Also set the ao pointer in the mixer to NULL
This avoids the problem addressed in the previous commit (r38414) by
making sure we don't call into AOs that were uninitialized.

Unfortunately it's not sufficient and causes a regression for AOs
where the call (even when the AO was already uninitialized) would
still work and unmute the audio (e.g. ao alsa).

So for example if a file is played with ao alsa and exits in muted
state. The mute will not be reversed and the alsa output device will
stay muted. If this is the user's main audio output device it would
usually mean the audio is completely muted for her system because
alsa is usually **the** low level audio device on her Linux system
used by everything else as well.

Patch by Reimar



git-svn-id: svn://svn.mplayerhq.hu/mplayer/trunk@38415 b3059339-0415-0410-9bf9-f77b7e298cf2