mplayer: Do not un-initialize components before exit_player_with_rc
After we finished all play tree items, we will call exit_player_with_rc.

When calling exit_player_with_rc, internally the routine will try
to uninit everything by itself.

More important: This way we unmute the AO if it's muted.

This avoids the regression introduced by the previous commit (r38415)
for typical usage, but it won't help for all potential cases. We still
fail to unmute correctly if one is using multiple AOs like this:

    mplayer file1 -ao pulse file2 -ao alsa file3 -ao pulse

It can even bite with one AO if there is a change to a file without
audio in as the last play tree item:

    ./mplayer file1 file2-without-audio-streams



git-svn-id: svn://svn.mplayerhq.hu/mplayer/trunk@38416 b3059339-0415-0410-9bf9-f77b7e298cf2