fmt-conversion: Remove reference to AV_PIX_FMT_XVMC
The pix fmt was removed from lavu on major bump to 59.

This fixes the build when vo xvmc is not enabled in configure.

As vo xvmc isn't automatically enabled, this commit should be
enough to avoid build breakage with newer FFmpeg for most people.
Even those that have the relevant dev packages installed.

More stuff related to XvMC (vo xvmc, parts in vd ffmpeg, etc) are
planned to be removed in follow up commits.



git-svn-id: svn://svn.mplayerhq.hu/mplayer/trunk@38450 b3059339-0415-0410-9bf9-f77b7e298cf2