ao sdl: Switch from AVFifoBuffer to AVFifo
Removed from lavu on major bump to 59

Switched to the version with SDL_{Lock,Unlock}Audio.

That's probably not in the spirit of the original version
of ao sdl, which tried hard to avoid locking, but as the
new version of FFmpeg fifo buffer seems to be worse for
multi threaded use I rather prefer to be safe here.

P.S.
I didn't really look deep into the details.
The old version of the fifo buffer probably wasn't completely
safe (especially on all CPU archs), but from the code history
it at least tried to.



git-svn-id: svn://svn.mplayerhq.hu/mplayer/trunk@38455 b3059339-0415-0410-9bf9-f77b7e298cf2