Limit the number of files returned by the playtree builder.
The option "-loop 0 <URL|file>" will return an infinite number of
URLs or files, so limit it.

Also, redraw the GUI (and process its waiting events) when playback
ends. If a stream fails to open and play, there will be no GUI_REDRAW
otherwise, which would render the GUI unresponsive.

This fixes https://bugs.debian.org/418855.


git-svn-id: svn://svn.mplayerhq.hu/mplayer/trunk@38510 b3059339-0415-0410-9bf9-f77b7e298cf2