Keep old auto_quality value when canceling the preferences dialog.
Check its current value and reset it if necessary when canceling
the dialog.

Also, do not set it when pressing OK, because prHScaleChanged()
will have already done so if it has been changed.

Moreover, set its initial value when creating the dialog, and check
the value passed to mplayer()'s MPLAYER_SET_AUTO_QUALITY.


git-svn-id: svn://svn.mplayerhq.hu/mplayer/trunk@38619 b3059339-0415-0410-9bf9-f77b7e298cf2