Fix lost full screen mode and lost video menu items.
This all disappeared in svn r38514, where the video driver for the GUI
was no longer implicitly set, and without a specified video driver,
GUI_PREPARE could not decide if there would be a video window.

This is now determined later at GUI_SET_VIDEO when a video output driver
is finally set.


git-svn-id: svn://svn.mplayerhq.hu/mplayer/trunk@38638 b3059339-0415-0410-9bf9-f77b7e298cf2