configure: Make it possible to switch to a different branch in a shallow ffmpeg clone
If ffmpeg is automatically checked out by configure, we end up
with a shallow clone by default. That's good because it saves
time and bandwith.

It has the drawback that a later request to switch to a different
branch will fail.

Shallow fetch the dependent objects for the requested branch to
make it possible to switch to the requested branch even if it
differs from the initially requested branch.



git-svn-id: svn://svn.mplayerhq.hu/mplayer/trunk@38678 b3059339-0415-0410-9bf9-f77b7e298cf2