Workaround vo_vdpau not working on Radeon with libvdpau_va_gl VA-API wrapper.
Mesa 25.3.0 has removed native support for VDPAU. Still it could be used
with the "libvdpau_va_gl" wrapper library that translates it to VA-API.

However the Mesa VA-API driver for Radeon cards does a minimum size check.
Apparently the hardware limit for decoding is 64x64 resolution for everything
except VP9 and AV1 (they are at 16x16).

This causes a problem because vo_vdpau() tries to create a (dummy) decoder
with 48x48 resolution when query_format() check is executed.
Failing the check prevents further accelerated playback.

Workaround the problem by increasing the query_format() resolution to 64x64.


git-svn-id: svn://svn.mplayerhq.hu/mplayer/trunk@38681 b3059339-0415-0410-9bf9-f77b7e298cf2